continue lib api cleanup

+++ libfreeipmi/include/freeipmi/util/ipmi-rmcpplus-util.h
@@ -143,7 +143,7 @@ int ipmi_rmcpplus_check_remote_console_session_id (fiid_obj_t obj_cmd,
 int ipmi_rmcpplus_check_session_id (fiid_obj_t obj_rmcpplus_session_hdr,
                                     uint32_t session_id);
 
-int8_t ipmi_rmcpplus_calculate_payload_type (uint8_t *pkt, uint32_t pkt_len);
+int ipmi_rmcpplus_calculate_payload_type (uint8_t *pkt, uint32_t pkt_len);
 
 #ifdef __cplusplus
 }
