+++ libfreeipmi/src/debug/ipmi-debug-common.h
@@ -30,11 +30,11 @@
 /* Portable version of the extremely unportable Linux dprintf() */
 int ipmi_debug_dprintf (int fd, char *fmt, ...);
 
-int8_t ipmi_debug_set_prefix (char *buf, unsigned int buflen, const char *prefix);
+int ipmi_debug_set_prefix (char *buf, unsigned int buflen, const char *prefix);
 
-int8_t ipmi_debug_output_str (int fd, const char *prefix, const char *str);
+int ipmi_debug_output_str (int fd, const char *prefix, const char *str);
 
-int8_t ipmi_debug_output_byte_array (int fd, const char *prefix, uint8_t *buf, uint32_t buf_len);
+int ipmi_debug_output_byte_array (int fd, const char *prefix, uint8_t *buf, uint32_t buf_len);
 
 #endif /* ipmi-debug-common.h */
